turned photocarousel into reusable code
