+++ src/components/views/ApplicationViews.js
@@ -1,6 +1,6 @@
 import { NavBar } from "../nav/NavBar"
 import { Landing } from "../landing/Landing"
-import { ParkPage } from "../landing/ParkPage"
+import { ParkPage } from "../parks/ParkPage"
 import { Route, Outlet, Routes } from "react-router-dom"
 import { BlogContainer } from "../blog/BlogContainer"
 
@@ -17,7 +17,6 @@ export const ApplicationViews = () => {
 					<BlogContainer />
 					</>
 				} />
-				<Route path="/" element={ <Landing /> } />
 				<Route path="home" element={ <Landing /> } />
 				<Route path="parks/:park_id" element={ <ParkPage /> } />
 			</Route>
